+++ lib/app/di/service_locator.config.dart
@@ -13,6 +13,8 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart' as _i558;
 import 'package:get_it/get_it.dart' as _i174;
 import 'package:injectable/injectable.dart' as _i526;
+import 'package:jellyfinity/app/connected_playback/ActiveRemotePlaybackWatcher.dart'
+    as _i602;
 import 'package:jellyfinity/app/connected_playback/ConnectedPlaybackLink.dart'
     as _i419;
 import 'package:jellyfinity/app/connected_playback/ConnectedPlaybackTargetLink.dart'
@@ -669,6 +671,14 @@ extension GetItInjectableX on _i174.GetIt {
         gh<_i809.SessionCubit>(),
         gh<_i126.PlaybackCubit>(),
         gh<_i217.ConnectedPlaybackTargetLink>(),
+        gh<_i846.PlaybackControlCubit>(),
+      ),
+    );
+    gh.lazySingleton<_i602.ActiveRemotePlaybackWatcher>(
+      () => _i602.ActiveRemotePlaybackWatcher(
+        gh<_i40.DevicePresenceSource>(),
+        gh<_i809.SessionCubit>(),
+        gh<_i126.PlaybackCubit>(),
         gh<_i846.PlaybackControlCubit>(),
       ),
     );